Initial building of containers from Maven, see #5292
This commit adds initial support for building an "application only" container.
That is achived by adding the fabric8io/docker-maven-plugin to a new
Maven profile called "container". Build via `mvn docker:build`.

The build is based on the upstream Payara 5 image right now. This is a TODO,
as the upstream container project is not very responsive and there is no
proper signal handling in place.