Create JDBC connection from glassfish-resources.xml
This commit is a first test to apply database configuration and
credentials to the application server without using a config
on domain level but on application level.

Most certainly, this is only compatible with Payara 5.x as
Glassfish 5.x is unknown to support system properties in
config files and it certainly does not support password aliases.

The JNDI names have been changed to be conform with Java EE 7.
See https://github.com/javaee-samples/javaee7-samples/tree/master/jpa/datasourcedefinition-applicationxml-pu
and others for working examples. (Staying with the old name
was not successfull.)