feat,doc(setting): fixate that old names cannot change order of placeholders
For the sake of keeping semantics intact (and make it much easier to search for config bits),
no longer allow to use positional string replacements. As we are using String.format() in
AliasConfigSource, this was possible, but it doesn't make much sense.

Usually, the placeholders are going to be used to express some kind of object or a list of options.
Changing the order of that seems unlikely to be useful, as it would invert the semantic.

Also straightened the docs on old names some more, including current limitations of 'em.